requirements.txt
Flask
Flask-SQLAlchemy
mysql-python
pika
docker-py
json
requests
